Implement hot-reloading the server on a app bundle change
Previously, the server would only restart when it's directly referenced files
where built. An external build step was needed to create the bunlde. Now the
server creates it's own webpack compiler that will watch for file changes and
recreate the bundle. Once that happens, the serer will be recreated because of
the dependency on the bundle itself.

One more piece missing: client-side HMR.